src/backend/requirements.txt
# Framework
fastapi==0.109.2
uvicorn==0.27.1
Flask==2.2.3
Flask-RESTful==0.3.9
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Werkzeug==2.2.3
Jinja2==3.1.2
itsdangerous==2.1.2
MarkupSafe==2.1.2

# Database
psycopg2-binary==2.9.5
pymongo==4.3.3
redis==4.5.1
fakeredis==2.10.2

# Data Processing
pandas==1.5.3
numpy==1.24.2
scikit-learn==1.2.2

# Utils
python-dotenv==1.0.0
gunicorn==20.1.0
requests==2.28.2
pyjwt==2.6.0
marshmallow==3.19.0
celery==5.2.7
click==8.1.3
pytz==2022.7.1
python-dateutil==2.8.2
bcrypt==4.0.1
passlib==1.7.4
jsonschema==4.17.3

# Testing
pytest==7.2.2
pytest-cov==4.0.0

# Development
black==23.1.0
flake8==6.0.0
isort==5.12.0 